Configured Seeds.rb to aid testing code
This commit is VERY large (in terms of file size), it contains multiple
media files and images that are used by seeds.rb to insert test data
into the database.
The aim here is to ensure that while development continues, we're
working with unified test data before continuing onto the primary
testing phase.

Files are not intended to be content-indicative of the site's final
form.

Downloading this commit will not automatically overwrite your own
database.
To use this commit correctly, issue the command 'rake db:reset' from the
terminal at the rails root folder.